Fix fopen extended mode support and add offset checking to fseek
The extended '+' flag for fopen wasn't being parsed and the file mode
used to open files was wrong making it impossible to open a file for
both reading and writing.

TODO: Fix 'a+' file mode so that freads start at the start and fwrites
always happen at the end.

When seeking from SEEK_CUR, fseek only cared if lseek returned a
positive offset. I added a check to make sure that the given offset and
current position added together was the same position returned by lseek.