Add fma() from FreeBSD with some fixes for nonstandard code.
The macros in fenv.h are implementation defined, according to the
standard, and should be checked prior to use to tailor code for
different environments.

The fma() function passes the testsuite while simulating the r5900
floating point environment. But, the r5900 uses emulated double
precision support so I'm not sure if any additional changes are needed.